Add example for step by step
